Make `meta` work like `attribute`
All in all it makes `meta` behave exactly the same a `attribute`,
because in the end they both do the same thing: output a hash based on
the record and params.

Key changes, now possible with `meta` like with `attribute` already:

- `meta` accepts a parameter now. E.g. to output `meta: { some: :key }`,
  it is possible to do `meta :some { :key }` now.
- It can be conditional with `if: proc {}`: `meta :some, if: proc { ... }`
- Multiple `meta` calls are possible now, each adding a key to the meta
  hash
- Without a block it calls record's attribute with that name:
  `meta :attr` will call `record.attr`. This might not be very useful for
  `meta` but to keep the code the same it works like this and shouldn't
  hurt.

To stay backwards compatible it is still possible to call `meta` with
only a block and it will do the same as it did before. Maybe this
should be deprecated in the long term.

(cherry picked from commit 1c92e4102c4536b48c82ad8926e6073e0fd4c2ac)